CEDS-2178 Organise views and corresponding specs into packages
