+++ app/controllers/consolidations/DisassociateUCRController.scala
@@ -19,12 +19,12 @@ package controllers.consolidations
 import controllers.actions.{AuthenticatedAction, JourneyRefiner}
 import forms.DisassociateUcr
 import javax.inject.{Inject, Singleton}
-import models.cache.{Cache, DisassociateUcrAnswers, JourneyType}
+import models.cache.{DisassociateUcrAnswers, JourneyType}
 import play.api.i18n.I18nSupport
 import play.api.mvc._
 import repositories.CacheRepository
 import uk.gov.hmrc.play.bootstrap.controller.FrontendController
-import views.html.disassociate_ucr
+import views.html.disassociateucr.disassociate_ucr
 
 import scala.concurrent.{ExecutionContext, Future}
