+++ app/views/shutmucr/shut_mucr.scala.html
