+++ test/unit/controllers/consolidations/AssociateUCRControllerSpec.scala
@@ -26,7 +26,7 @@ import play.api.libs.json.Json
 import play.api.test.Helpers._
 import play.twirl.api.HtmlFormat
 import services.MockCache
-import views.html.associate_ucr
+import views.html.associateucr.associate_ucr
 
 import scala.concurrent.ExecutionContext.global
