+++ test/unit/controllers/consolidations/DisassociateUCRControllerSpec.scala
@@ -26,7 +26,7 @@ import play.api.libs.json.Json
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import services.MockCache
-import views.html.disassociate_ucr
+import views.html.disassociateucr.disassociate_ucr
 
 import scala.concurrent.ExecutionContext.Implicits.global
