+++ test/unit/controllers/consolidations/DisassociateUCRSummaryControllerSpec.scala
@@ -31,7 +31,7 @@ import play.api.http.Status
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import services.SubmissionService
-import views.html.disassociate_ucr_summary
+import views.html.disassociateucr.disassociate_ucr_summary
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
