+++ test/unit/controllers/consolidations/MucrOptionsControllerSpec.scala
@@ -28,7 +28,7 @@ import play.api.libs.json.{JsString, Json}
 import play.api.test.Helpers._
 import play.twirl.api.HtmlFormat
 import services.MockCache
-import views.html.mucr_options
+import views.html.associateucr.mucr_options
 
 import scala.concurrent.ExecutionContext.global
